# DOM renderer: Don't output empty cells at end of a line