Fixes #1609

---

This is an optimization to help make the DOM renderer more viable if the canvas renderer is replaced by https://github.com/xtermjs/xterm.js/pull/1790

---

Using the testing process outlined in https://github.com/xtermjs/xterm.js/wiki/Performance-testing#renderer (numbers measured on dc077a1)

**87x26 viewport**

Before: frames 16, average 11.53ms
After: frames 24, average 5.88ms

**300x80 viewport**

Before: frames 11, average 115.83ms
After: frames 13, average 17.12ms