@Tyriar 2018-11-23
This row overflow bug was fixed in https://github.com/xtermjs/xterm.js/pull/1792/commits/478dfee6e2988fd81a61b4e85b14125b3e087e50

<img width="552" alt="screen shot 2018-11-23 at 6 50 53 am" src="https://user-images.githubusercontent.com/2193314/48949273-44470e00-eeec-11e8-892b-7890192c7b4f.png">